requirements.txt
Django>=4.2
djangorestframework
django-admin-interface
django-colorfield
django-simple-history
psycopg2-binary
pandas
python-dotenv
celery
requests
python-jose
pyjwt[crypto]
